Noticed something with money, fixed it

+++ src/main/java/io/github/teambanhammer/configuration.java
@@ -20,7 +20,7 @@ public class configuration {
         Config.addDefault("teleportTimer",3);
         Config.addDefault("requestLifetime",60);
         Config.addDefault("useVault" , false);
-        Config.addDefault("teleportPrice" , 100);
+        Config.addDefault("teleportPrice" , 100.00);
         Config.addDefault("eventBeforeTP" , "&aTeleporting in &b{countdown} seconds&a, please don't move!");
         Config.addDefault("eventTeleport" , "&aTeleporting...");
         Config.addDefault("eventMovement" , "&cTeleport has been cancelled due to movement.");
@@ -46,7 +46,7 @@ public class configuration {
         return Config.getInt("requestLifetime");
     }
     public static boolean useVault() {return Config.getBoolean("useVault");}
-    public static int teleportPrice() {return Config.getInt("teleportPrice");}
+    public static double teleportPrice() {return Config.getDouble("teleportPrice");}
     public static String eventMovement() {return ChatColor.translateAlternateColorCodes('&' , Config.getString("eventMovement"));}
     public static String eventTeleport() {return ChatColor.translateAlternateColorCodes('&' , Config.getString("eventTeleport"));}
     public static String eventBeforeTP() {return ChatColor.translateAlternateColorCodes('&' , Config.getString("eventBeforeTP"));}
